Employee basic info fields reusability adjusted with pro
