+++ classes/Controllers/EmployeeController.php
@@ -19,7 +19,7 @@ class EmployeeController {
 			'role' => 'administrator',
 		),
 		'fetchEmployee'          => array(
-			'role' => 'administrator',
+			'nopriv' => true,
 		),
 		'getEmployeeList'        => array(
 			'role' => 'administrator',
@@ -94,6 +94,11 @@ class EmployeeController {
 	 * @return void
 	 */
 	public static function fetchEmployee( int $user_id ) {
+
+		// Validate access
+		if ( get_current_user_id() != $user_id && ! User::validateRole( get_current_user_id(), apply_filters( 'crewhrm_hr_roles', array( 'administrator' ) ) ) ) {
+			wp_send_json_error( array( 'message' => __( 'Access denied!', 'crewhrm' ) ) );
+		}
 
 		$employee = User::getUserInfo( $user_id );
